renamed the plugin with packaging handled
