--- src/piopiy/services/opensource/chatterbox/__init__.py
@@ -1,10 +1,0 @@
-# piopiy/services/opensource/kokoro/__init__.py
-from .tts_client import *
-from .tts import *   # load KokoroTTSService and others into namespace
-from .mtl_tts import *
-
-import sys
-from piopiy.services import DeprecatedModuleProxy
-
-# Wrap this module in a proxy so you can warn about deprecated paths
-sys.modules[__name__] = DeprecatedModuleProxy(globals(), "chatterbox", "chatterbox.tts")
